Rename `SortMenuV2` to `SortMenu`
