+++ web/packages/teleport/src/Integrations/Enroll/Shared/FilterPanel.tsx
@@ -25,7 +25,7 @@ import {
   SortItem,
   SortMenu,
   SortOrder,
-} from 'shared/components/Controls/SortMenuV2';
+} from 'shared/components/Controls/SortMenu';
 
 import {
   integrationTagOptions,
